requirements.txt
aiogram==2.22
aiogram-calendar==0.1.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
flake8==5.0.4
frozenlist==1.3.1
idna==3.3
loguru==0.6.0
mccabe==0.7.0
multidict==6.0.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot-calendar==1.0.5
pytz==2022.2.1
requests==2.28.1
six==1.16.0
urllib3==1.26.11
yarl==1.8.1
